Implement #92 맵 위치 이동에 따른 리스너 이벤트 추가

+++ app/src/main/java/com/example/bikini_android/network/request/param/NearbyFeedParameter.kt
@@ -18,7 +18,7 @@ class NearbyFeedParameter : RestApiParameter() {
         put(RestApiParameterKey.LATITUDE, value.latitude.toString())
         put(RestApiParameterKey.LONGITUDE, value.longitude.toString())
     }
-    fun setRadius(value: Double) {
+    fun setRadius(value: Float) {
         put(RestApiParameterKey.RADIUS, value.toString())
     }
