+++ app/src/main/java/com/example/bikini_android/repository/feed/FeedRepository.kt
@@ -27,6 +27,6 @@ interface FeedRepository {
 
     fun getNearbyFeedsFromRemote(
         latLng: LatLng,
-        radius: Double
+        radius: Float
     ): Single<List<Feed>>
 }
